Код непосредственно сортировок вынесен в отдельный модуль
